+++ src/main.ts
@@ -1,4 +1,5 @@
 import "./types/theme"
-import theme from "./light"
 
-export default theme
+export { default } from "./light"
+export { default as themeDark } from "./dark"
+export { default as themeCombined } from "./themeCombined"
